Extract marker code from scene editor into marker tracker
