+++ core/ide/src/main/java/org/alice/stageide/sceneeditor/StorytellingSceneEditor.java
@@ -99,7 +99,6 @@ import org.alice.stageide.run.RunComposite;
 import org.alice.stageide.sceneeditor.draganddrop.SceneDropSite;
 import org.alice.stageide.sceneeditor.interact.CameraNavigatorWidget;
 import org.alice.stageide.sceneeditor.interact.GlobalDragAdapter;
-import org.alice.stageide.sceneeditor.interact.manipulators.OrthographicCameraDragZoomManipulator;
 import org.alice.stageide.sceneeditor.side.SideComposite;
 import org.alice.stageide.sceneeditor.snap.SnapState;
 import org.alice.stageide.sceneeditor.viewmanager.CameraMarkerTracker;
@@ -338,25 +337,11 @@ public class StorytellingSceneEditor extends AbstractSceneEditor implements Rend
   private final Button runButton = IsToolBarShowing.getValue() ? null : RunComposite.getInstance().getLaunchOperation().createButton();
 
   private OrthographicCameraImp orthographicCameraImp = null;
-  private OrthographicCameraMarkerImp topOrthoMarkerImp = null;
-  private OrthographicCameraMarkerImp frontOrthoMarkerImp = null;
-  private OrthographicCameraMarkerImp sideOrthoMarkerImp = null;
   private PerspectiveCameraMarkerImp startingCameraMarkerImp;
-  private PerspectiveCameraMarkerImp layoutSceneMarkerImp;
 
   private ComboBox<CameraOption> mainCameraViewSelector;
   private CameraMarkerTracker mainCameraViewTracker;
   private CameraOption savedSceneEditorViewSelection = null;
-
-  public static int STARTING_CAMERA = 0;
-  public static int LAYOUT_CAMERA = 1;
-  private final double DEFAULT_LAYOUT_CAMERA_Y_OFFSET = 12.0;
-  private final double DEFAULT_LAYOUT_CAMERA_Z_OFFSET = 10.0;
-  private final double DEFAULT_TOP_CAMERA_Y_OFFSET = 10.0;
-  private final double DEFAULT_SIDE_CAMERA_X_OFFSET = 10.0;
-  private final double DEFAULT_FRONT_CAMERA_Z_OFFSET = 10.0;
-  private final int DEFAULT_LAYOUT_CAMERA_ANGLE = -40;
-
 
   private ImmutableDataSingleSelectListState<CameraOption> mainCameraMarkerList = ViewListSelectionState.getInstance();
 
@@ -370,7 +355,7 @@ public class StorytellingSceneEditor extends AbstractSceneEditor implements Rend
   }
 
   public void setStartingCameraMarkerTransformation(AffineMatrix4x4 transform) {
-    startingCameraMarkerImp.setLocalTransformation(transform);
+    mainCameraViewTracker.setStartingCameraMarkerTransformation(transform);
   }
 
   public static class SceneEditorProgramImp extends ProgramImp {
@@ -452,103 +437,8 @@ public class StorytellingSceneEditor extends AbstractSceneEditor implements Rend
     }
   }
 
-  private double clampCameraValue(double val) {
-    final double CAMERA_CLAMP_MAX = 100.0;
-    final double CAMERA_CLAMP_MIN = 2.0;
-    return clampValue(val, CAMERA_CLAMP_MIN, CAMERA_CLAMP_MAX);
-  }
-
-  private double clampPictureValue(double val) {
-    // we intentionally allow a slightly larger max here, and the manipulator deals with it later.
-    final double PICTURE_CLAMP_MAX = 100;
-    final double PICTURE_CLAMP_MIN = OrthographicCameraDragZoomManipulator.MIN_ZOOM;
-    return clampValue(val, PICTURE_CLAMP_MIN, PICTURE_CLAMP_MAX);
-  }
-
-  private double clampValue(double val, double min, double max) {
-    return Math.min(max, Math.max(min, val));
-  }
-
-
   public void centerCameraOn(UserField field) {
-    // Changes the markers for the layout camera and the 3 ortho camera views, not just whatever marker is active.
-    Object instanceInJava = IDE.getActiveInstance().getSceneEditor().getInstanceInJavaVMForField(field);
-    EntityImp target = EmployeesOnly.getImplementation((SThing) instanceInJava);
-    AxisAlignedBox alignedBox = target.getDynamicAxisAlignedMinimumBoundingBox(org.lgna.story.implementation.AsSeenBy.SCENE);
-    double targetHeight = alignedBox.getHeight();
-    double targetWidth = alignedBox.getWidth();
-    double targetDepth = alignedBox.getDepth();
-    Point3 targetTranslation = alignedBox.getCenter();
-
-    // Update layout camera marker
-    // Attempting to find the least disruptive move, ideally a bit above of the object (because looking up through the
-    // object/ground feels bad) and far enough away that we can see most of it, but hopefully close enough that there
-    // isn't another object in between. Usually we succeed.
-
-    // Since targetTranslation is already centered in y, this is effectively height * 1.5.
-    double adjustedY = targetTranslation.y + Math.max(targetHeight, alignedBox.getDiagonal() * .5);
-    Point3 adjustedPos = new Point3(targetTranslation.x,  adjustedY, targetTranslation.z);
-
-    // if the camera is already above it, great. Otherwise we get the best results by using the same adjusted y.
-    Point3 adjustedCameraPos = layoutSceneMarkerImp.getAbsoluteTransformation().translation;
-    adjustedCameraPos.y = Math.max(adjustedCameraPos.y, adjustedY);
-
-    Vector3 direction = Vector3.createSubtraction(adjustedCameraPos, adjustedPos);
-    direction.normalize();
-    Ray ray = new Ray(adjustedPos, direction);
-    Point3 layoutCamTranslation = ray.getPointAlong(alignedBox.getDiagonal() * 1.5);
-
-    // orientation calculated from wherever our camera ended up to look at the center of the object.
-    Vector3 cameraDirection = Vector3.createSubtraction(layoutCamTranslation, targetTranslation);
-    cameraDirection.normalize();
-    ForwardAndUpGuide forwardAndUpGuide = new ForwardAndUpGuide(Vector3.createNegation(cameraDirection), null);
-    OrthogonalMatrix3x3 layoutCamOrientation = forwardAndUpGuide.createOrthogonalMatrix3x3();
-
-    AffineMatrix4x4 layoutTransform = AffineMatrix4x4.createIdentity();
-    layoutTransform.applyTranslation(layoutCamTranslation);
-    layoutTransform.applyOrientation(layoutCamOrientation);
-    layoutSceneMarkerImp.setLocalTransformation(layoutTransform);
-
-    // Update Orthographic camera markers
-    // Translation helps clip things that might be over our object, based on how large it is. (It doesn't always succeed)
-    // PicturePlane controls how much of the scene is in our view, aka 'zoom'
-    ClippedZPlane picturePlane = new ClippedZPlane();
-
-    // Top
-    AffineMatrix4x4 topTransform = AffineMatrix4x4.createIdentity();
-    topTransform.translation.x = targetTranslation.x;
-    topTransform.translation.y = targetTranslation.y + (targetHeight != 0 ? clampCameraValue(targetHeight * DEFAULT_TOP_CAMERA_Y_OFFSET) : DEFAULT_TOP_CAMERA_Y_OFFSET);
-    topTransform.translation.z = targetTranslation.z;
-    topTransform.orientation.up.set(0, 0, 1);
-    topTransform.orientation.right.set(-1, 0, 0);
-    topTransform.orientation.backward.set(0, 1, 0);
-    topOrthoMarkerImp.setLocalTransformation(topTransform);
-    picturePlane.setCenter(0, 0);
-    picturePlane.setHeight(clampPictureValue(Math.max(targetDepth, RunComposite.WIDTH_TO_HEIGHT_RATIO * targetWidth)));
-    topOrthoMarkerImp.setPicturePlane(picturePlane);
-
-    // Side
-    AffineMatrix4x4 sideTransform = AffineMatrix4x4.createIdentity();
-    sideTransform.translation.x = targetTranslation.x + (targetWidth != 0 ? clampCameraValue(targetWidth * DEFAULT_SIDE_CAMERA_X_OFFSET) : DEFAULT_SIDE_CAMERA_X_OFFSET);
-    sideTransform.translation.y = targetTranslation.y;
-    sideTransform.translation.z = targetTranslation.z;
-    sideTransform.orientation.setValue(new ForwardAndUpGuide(Vector3.accessNegativeXAxis(), Vector3.accessPositiveYAxis()));
-    sideOrthoMarkerImp.setLocalTransformation(sideTransform);
-    picturePlane.setHeight(clampPictureValue(Math.max(targetDepth, RunComposite.WIDTH_TO_HEIGHT_RATIO * targetHeight)));
-    sideOrthoMarkerImp.setPicturePlane(picturePlane);
-
-    // Front
-    AffineMatrix4x4 frontTransform = AffineMatrix4x4.createIdentity();
-    frontTransform.translation.x = targetTranslation.x;
-    frontTransform.translation.y = targetTranslation.y;
-    frontTransform.translation.z = targetTranslation.z - (targetDepth != 0 ? clampCameraValue(targetDepth * DEFAULT_FRONT_CAMERA_Z_OFFSET) : DEFAULT_FRONT_CAMERA_Z_OFFSET);
-    frontTransform.orientation.setValue(new ForwardAndUpGuide(Vector3.accessPositiveZAxis(), Vector3.accessPositiveYAxis()));
-    frontOrthoMarkerImp.setLocalTransformation(frontTransform);
-    picturePlane.setHeight(clampPictureValue(Math.max(targetWidth, RunComposite.WIDTH_TO_HEIGHT_RATIO * targetHeight)));
-    frontOrthoMarkerImp.setPicturePlane(picturePlane);
-
-    // Update camera to the latest markers
-    mainCameraViewTracker.updateCameraToNewMarkerLocation();
+    mainCameraViewTracker.centerMarkersOn(field);
   }
 
   @Override
@@ -599,99 +489,17 @@ public class StorytellingSceneEditor extends AbstractSceneEditor implements Rend
             && !valueType.isAssignableFrom(SVRHand.class);
   }
 
-  private void initializeCameraMarkers() {
-    PerspectiveCameraMarker openingSceneMarker = new PerspectiveCameraMarker();
-    openingSceneMarker.setColorId(Color.DARK_GRAY);
-    this.startingCameraMarkerImp = EmployeesOnly.getImplementation(openingSceneMarker);
-    this.startingCameraMarkerImp.setDisplayVisuals(true);
-    this.startingCameraMarkerImp.setCameraType(STARTING_CAMERA);
-    startingCameraMarkerImp.setVrActive(isVrActive());
-    MarkerUtilities.addIconForCameraImp(this.startingCameraMarkerImp, "mainCamera");
-    MarkerUtilities.setViewForCameraImp(this.startingCameraMarkerImp, CameraOption.STARTING_CAMERA_VIEW);
-
-    PerspectiveCameraMarker layoutCameraMarker = new PerspectiveCameraMarker();
-    layoutCameraMarker.setColorId(Color.LIGHT_BLUE);
-    this.layoutSceneMarkerImp = EmployeesOnly.getImplementation(layoutCameraMarker);
-    this.layoutSceneMarkerImp.setDisplayVisuals(true);
-    this.layoutSceneMarkerImp.setCameraType(LAYOUT_CAMERA);
-    layoutSceneMarkerImp.setVrActive(isVrActive());
-    MarkerUtilities.addIconForCameraImp(this.layoutSceneMarkerImp, "sceneEditorCamera");
-    MarkerUtilities.setViewForCameraImp(this.layoutSceneMarkerImp, CameraOption.LAYOUT_SCENE_VIEW);
-
-    OrthographicCameraMarker topOrthoMarker = new OrthographicCameraMarker();
-    this.topOrthoMarkerImp = EmployeesOnly.getImplementation(topOrthoMarker);
-    MarkerUtilities.addIconForCameraImp(this.topOrthoMarkerImp, "top");
-    MarkerUtilities.setViewForCameraImp(this.topOrthoMarkerImp, CameraOption.TOP);
-    AffineMatrix4x4 topTransform = AffineMatrix4x4.createIdentity();
-    topTransform.translation.y = 10;
-    topTransform.translation.z = -10;
-    topTransform.orientation.up.set(0, 0, 1);
-    topTransform.orientation.right.set(-1, 0, 0);
-    topTransform.orientation.backward.set(0, 1, 0);
-    assert topTransform.orientation.isWithinReasonableEpsilonOfUnitLengthSquared();
-    this.topOrthoMarkerImp.setLocalTransformation(topTransform);
-    ClippedZPlane picturePlane = new ClippedZPlane();
-    picturePlane.setCenter(0, 0);
-    picturePlane.setHeight(16);
-    this.topOrthoMarkerImp.setPicturePlane(picturePlane);
-
-    OrthographicCameraMarker sideOrthoMarker = new OrthographicCameraMarker();
-    this.sideOrthoMarkerImp = EmployeesOnly.getImplementation(sideOrthoMarker);
-    MarkerUtilities.addIconForCameraImp(this.sideOrthoMarkerImp, "side");
-    MarkerUtilities.setViewForCameraImp(this.sideOrthoMarkerImp, CameraOption.SIDE);
-    AffineMatrix4x4 sideTransform = AffineMatrix4x4.createIdentity();
-    sideTransform.translation.x = 10;
-    sideTransform.translation.y = 1;
-    sideTransform.orientation.setValue(new ForwardAndUpGuide(Vector3.accessNegativeXAxis(), Vector3.accessPositiveYAxis()));
-    assert sideTransform.orientation.isWithinReasonableEpsilonOfUnitLengthSquared();
-    this.sideOrthoMarkerImp.setLocalTransformation(sideTransform);
-    picturePlane.setHeight(4);
-    this.sideOrthoMarkerImp.setPicturePlane(picturePlane);
-
-    OrthographicCameraMarker frontOrthoMarker = new OrthographicCameraMarker();
-    this.frontOrthoMarkerImp = EmployeesOnly.getImplementation(frontOrthoMarker);
-    MarkerUtilities.addIconForCameraImp(this.frontOrthoMarkerImp, "front");
-    MarkerUtilities.setViewForCameraImp(this.frontOrthoMarkerImp, CameraOption.FRONT);
-    AffineMatrix4x4 frontTransform = AffineMatrix4x4.createIdentity();
-    frontTransform.translation.z = -10;
-    frontTransform.translation.y = 1;
-    frontTransform.orientation.setValue(new ForwardAndUpGuide(Vector3.accessPositiveZAxis(), Vector3.accessPositiveYAxis()));
-    assert frontTransform.orientation.isWithinReasonableEpsilonOfUnitLengthSquared();
-    this.frontOrthoMarkerImp.setLocalTransformation(frontTransform);
-    picturePlane.setHeight(4);
-    this.frontOrthoMarkerImp.setPicturePlane(picturePlane);
-
-    this.startingCameraMarkerImp.getAbstraction().setName(MarkerUtilities.getNameForCameraImp(this.startingCameraMarkerImp));
-    this.layoutSceneMarkerImp.getAbstraction().setName(MarkerUtilities.getNameForCameraImp(this.layoutSceneMarkerImp));
-    this.topOrthoMarkerImp.getAbstraction().setName(MarkerUtilities.getNameForCameraImp(this.topOrthoMarkerImp));
-    this.sideOrthoMarkerImp.getAbstraction().setName(MarkerUtilities.getNameForCameraImp(this.sideOrthoMarkerImp));
-    this.frontOrthoMarkerImp.getAbstraction().setName(MarkerUtilities.getNameForCameraImp(this.frontOrthoMarkerImp));
-  }
-
   public Boolean isVrActive() {
     return isVrScene;
   }
 
   private void setIsVrActive(boolean isActive) {
     isVrScene = isActive;
-    startingCameraMarkerImp.setVrActive(isVrScene);
-    layoutSceneMarkerImp.setVrActive(isVrScene);
-
-    // Update menu icon
-    MarkerUtilities.addIconForCameraImp(startingCameraMarkerImp, isVrScene ? "vrHeadset" : "mainCamera");
+    mainCameraViewTracker.setIsVrActive(isActive);
 
     // Update camera marker move buttons
     MoveActiveCameraToMarkerActionOperation.getInstance().updateDisplay();
     MoveMarkerToActiveCameraActionOperation.getInstance().updateDisplay();
-  }
-
-  private void clearCameras() {
-    this.snapGrid.stopTrackingCameras();
-    if (this.onscreenRenderTarget.getSgCameraCount() > 0) {
-      onscreenRenderTarget.clearSgCameras();
-    }
-    this.mainCameraViewTracker.setCameras(null, null);
-    this.globalDragAdapter.clearCameraViews();
   }
 
   private void setCameras(SymmetricPerspectiveCamera perspectiveCamera, OrthographicCamera orthographicCamera) {
@@ -881,8 +689,6 @@ public class StorytellingSceneEditor extends AbstractSceneEditor implements Rend
     this.orthographicCameraImp = new OrthographicCameraImp();
     this.orthographicCameraImp.getSgCamera().nearClippingPlaneDistance.setValue(.01d);
 
-    initializeCameraMarkers();
-
     this.globalDragAdapter.addSelectionListener(new SelectionListener() {
       @Override
       public void selecting(SelectionEvent e) {
@@ -908,14 +714,10 @@ public class StorytellingSceneEditor extends AbstractSceneEditor implements Rend
     SideComposite.getInstance().getObjectMarkersTab().getMarkerListState().addAndInvokeNewSchoolValueListener(this.objectMarkerFieldSelectionListener);
 
     this.mainCameraViewTracker = new CameraMarkerTracker(this, animator);
+    startingCameraMarkerImp = mainCameraViewTracker.startingCameraMarkerImp;
     this.mainCameraViewSelector = this.mainCameraMarkerList.getPrepModel().createComboBox();
     this.mainCameraViewSelector.setRenderer(new CameraViewCellRenderer(this.mainCameraViewTracker));
     this.mainCameraViewSelector.setFontSize(15);
-    this.mainCameraViewTracker.mapViewToMarker(CameraOption.STARTING_CAMERA_VIEW, this.startingCameraMarkerImp);
-    this.mainCameraViewTracker.mapViewToMarker(CameraOption.LAYOUT_SCENE_VIEW, this.layoutSceneMarkerImp);
-    this.mainCameraViewTracker.mapViewToMarker(CameraOption.TOP, this.topOrthoMarkerImp);
-    this.mainCameraViewTracker.mapViewToMarker(CameraOption.SIDE, this.sideOrthoMarkerImp);
-    this.mainCameraViewTracker.mapViewToMarker(CameraOption.FRONT, this.frontOrthoMarkerImp);
     this.mainCameraMarkerList.addAndInvokeNewSchoolValueListener(this.mainCameraViewTracker);
     this.mainCameraMarkerList.addAndInvokeNewSchoolValueListener(this.mainCameraViewSelectionObserver);
     this.lookingGlassPanel.addComponent(this.mainCameraViewSelector, Horizontal.CENTER, 0, Vertical.NORTH, 20);
@@ -1030,14 +832,8 @@ public class StorytellingSceneEditor extends AbstractSceneEditor implements Rend
         }
       }
 
-      AffineMatrix4x4 openingViewTransform = movableSceneCameraImp.getAbsoluteTransformation();
-      this.startingCameraMarkerImp.setLocalTransformation(openingViewTransform);
+      mainCameraViewTracker.updateLayoutCameraForScene(movableSceneCameraImp);
 
-      AffineMatrix4x4 layoutTransform = new AffineMatrix4x4(openingViewTransform);
-      layoutTransform.applyTranslationAlongYAxis(DEFAULT_LAYOUT_CAMERA_Y_OFFSET);
-      layoutTransform.applyTranslationAlongZAxis(DEFAULT_LAYOUT_CAMERA_Z_OFFSET);
-      layoutTransform.applyRotationAboutXAxis(new AngleInDegrees(DEFAULT_LAYOUT_CAMERA_ANGLE));
-      this.layoutSceneMarkerImp.setLocalTransformation(layoutTransform);
       savedSceneEditorViewSelection = null;
       this.mainCameraViewTracker.startTrackingCameraView(this.mainCameraMarkerList.getValue());
 
@@ -1460,7 +1256,7 @@ public class StorytellingSceneEditor extends AbstractSceneEditor implements Rend
   }
 
   public AffineMatrix4x4 getTransformForNewCameraMarker() {
-    return startingCameraMarkerImp.getAbsoluteTransformation();
+    return mainCameraViewTracker.getTransformForNewCameraMarker();
   }
 
   public AffineMatrix4x4 getTransformForNewObjectMarker() {
