+++ core/story-api/src/main/java/org/lgna/story/implementation/PerspectiveCameraMarkerImp.java
@@ -275,14 +275,6 @@ public class PerspectiveCameraMarkerImp extends CameraMarkerImp {
     }
   }
 
-  public void setCameraType(int newCameraType) {
-    this.cameraType = newCameraType;
-  }
-
-  public int getCameraType() {
-    return cameraType;
-  }
-
   @Override
   protected final SimpleAppearance[] getSgPaintAppearances() {
     return this.sgAppearances;
